requirements.txt
Django==5.2.7
Pillow==10.4.0
django-ckeditor==6.7.0
djangorestframework==3.15.2
PyMySQL==1.1.1